renamed and simplified
