--- components/Insights/index.jsx
@@ -1,22 +1,0 @@
-import Detail from './Detail';
-
-export default function Insights(props) {
-  const { metric, stats } = props;
-
-  return (
-    <>
-      <div className='flex items-end p-4'>
-        <h1 className='flex-0 font-bold text-4xl text-stone-600'>
-          {metric.name}:
-        </h1>
-        <p className='flex-1 font-bold text-4xl text-stone-800 mx-2'>{stats.value}</p>
-      </div>
-      <div className='my-8' >
-        <Detail 
-          metric={metric}
-          stats={stats}
-        />
-      </div>
-    </>
-  )
-}
